Assignment4/Chinese-Song-Tone-Analyzer/requirements.txt
blinker==1.4
Bootstrap-Flask==1.0.10
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
facebook-sdk==2.0.0
Flask==1.0.2
Flask-Login==0.3.2
Flask-Stormpath==0.4.8
Flask-WTF==0.14.2
google-api-core==1.8.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-cloud-translate==1.3.3
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
oauth2client==1.5.2
oauthlib==1.0.3
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyDispatcher==2.0.5
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
stormpath==2.4.4
urllib3==1.24.1
watson-developer-cloud==2.10.0
websocket-client==0.48.0
Werkzeug==0.15.1
WTForms==2.2.1
